requirements.txt
Pillow==8.4.0
python-telegram-bot==13.7
astral==2.2
pyttsx3==2.90
pytz~=2021.3
torch==1.9.1
transformers==4.11.3
numpy==1.21.2